pyproject.toml
[project]
name = "rag-chatbot"
version = "1.0.0"
description = "A Retrieval-Augmented Generation chatbot using Cohere embeddings, Qdrant vector store, and Gemini 2.5 Flash"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "python-dotenv>=1.0.0",
    "cohere>=5.0.0",
    "qdrant-client>=1.10.0",
    "google-generativeai>=0.8.0",
    "pydantic>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=24.0.0",
    "flake8>=6.0.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
